Add support for section headers
This commit adds support for PE section headers; namely, reading them
into memory and printing them to standard output. It should be noted
that the `Name` field in `struct Section_Header` is read as a sequence
of `uint8_t`s which are interpreted as `char`s when sent to standard
output. This is not technically in compliance with the Microsoft
specifications for the PE file format, as the `Name` field is stored in
PE files as UTF-8 encoded strings.